requirements.txt
torch
torchvision
torchaudio
opencv-python
pygame
numpy
pandas
matplotlib
scikit-learn
tqdm
pyyaml
einops
kaggle
